# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SUSE AI Universal Proxy provides RESTful APIs for managing MCP (Model Context Protocol) server deployments and proxying requests to MCP servers running in Kubernetes.",
        "title": "SUSE AI Universal Proxy - Control Plane",
        "contact": {
            "name": "SUSE AI Universal Proxy Team",
            "url": "https://github.com/SUSE/suse-ai-up"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/SUSE/suse-ai-up/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8911",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/adapters/{name}/client-token": {
            "get": {
                "description": "Retrieves a client token for authenticating with the MCP server adapter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adapters",
                    "authentication"
                ],
                "summary": "Get client token for adapter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adapter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ClientTokenResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/adapters/{name}/test-auth": {
            "post": {
                "description": "Tests the authentication configuration by attempting to connect to the MCP server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adapters",
                    "authentication"
                ],
                "summary": "Test adapter authentication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adapter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/adapters/{name}/validate-auth": {
            "post": {
                "description": "Validates the authentication configuration for an adapter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adapters",
                    "authentication"
                ],
                "summary": "Validate adapter authentication configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adapter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Authentication configuration to validate",
                        "name": "authConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AdapterAuthConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.APIKeyConfig": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "location": {
                    "description": "\"header\", \"query\", \"cookie\"",
                    "type": "string"
                },
                "name": {
                    "description": "Header name, query param, or cookie name",
                    "type": "string"
                }
            }
        },
        "models.AdapterAuthConfig": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "$ref": "#/definitions/models.APIKeyConfig"
                },
                "basic": {
                    "$ref": "#/definitions/models.BasicAuthConfig"
                },
                "bearerToken": {
                    "$ref": "#/definitions/models.BearerTokenConfig"
                },
                "oauth": {
                    "$ref": "#/definitions/models.OAuthConfig"
                },
                "required": {
                    "description": "true = require auth, false = optional",
                    "type": "boolean"
                },
                "token": {
                    "description": "Legacy field for backward compatibility",
                    "type": "string"
                },
                "type": {
                    "description": "\"bearer\", \"oauth\", \"basic\", \"apikey\", \"none\"",
                    "type": "string"
                }
            }
        },
...
        "models.BasicAuthConfig": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.BearerTokenConfig": {
            "type": "object",
            "properties": {
                "dynamic": {
                    "description": "Use token manager",
                    "type": "boolean"
                },
                "expiresAt": {
                    "type": "string"
                },
                "token": {
                    "description": "Static token",
                    "type": "string"
                }
            }
        },
...
        "models.OAuthConfig": {
            "type": "object",
            "properties": {
                "authUrl": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "clientSecret": {
                    "type": "string"
                },
                "redirectUri": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokenUrl": {
                    "type": "string"
                }
            }
        },
...
        "service.ClientTokenResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}